rename build to pipes
