+++ src/extensions/bit/bit.manifest.ts
@@ -2,13 +2,13 @@ import { WorkspaceExt } from '../workspace';
 import { ScopeExt } from '../scope';
 import { CapsuleExt } from '../capsule';
 import provideBit from './bit.provider';
-import { BuildExt } from '../build';
+import { PipesExt } from '../pipes';
 import { ComposerExt } from '../composer';
 import { ReactExtension } from '../react/react.manifest';
 
 export default {
   name: 'Bit',
-  dependencies: [WorkspaceExt, ScopeExt, CapsuleExt, BuildExt, ComposerExt, ReactExtension],
+  dependencies: [WorkspaceExt, ScopeExt, CapsuleExt, PipesExt, ComposerExt, ReactExtension],
   config: {},
   provider: provideBit
 };
