+++ src/extensions/composer/composer.manifest.ts
@@ -2,11 +2,11 @@ import { WatchExt } from '../watch';
 import { provideComposer } from './composer.provider';
 import { BitCliExt } from '../cli';
 import { WorkspaceExt } from '../workspace';
-import { BuildExt } from '../build';
+import { PipesExt } from '../pipes';
 
 export default {
   name: 'Composer',
-  dependencies: [WatchExt, BitCliExt, WorkspaceExt, BuildExt],
+  dependencies: [WatchExt, BitCliExt, WorkspaceExt, PipesExt],
   config: {},
   provider: provideComposer
 };
