+++ src/extensions/pipes/index.ts
@@ -1,0 +1,3 @@
+export { default as PipesExt } from './pipes.manifest';
+export { Pipes } from './pipes';
+export { TaskContext } from './task-context';
