+++ src/extensions/pipes/pipes.manifest.ts
@@ -1,10 +1,10 @@
-import { Build } from './build';
+import { Pipes } from './pipes';
 import { BitCliExt } from '../cli';
 import { WorkspaceExt } from '../workspace';
 import { CapsuleExt } from '../capsule';
 
 export default {
-  name: 'Build',
+  name: 'Pipes',
   dependencies: [BitCliExt, WorkspaceExt, CapsuleExt],
-  provider: Build.provide
+  provider: Pipes.provide
 };
