+++ src/extensions/react/react.manifest.ts
@@ -1,11 +1,11 @@
 import { ExtensionManifest } from '../../harmony';
-import { BuildExt, Build } from '../build';
+import { PipesExt, Pipes } from '../pipes';
 import { reactTask } from './react.task';
 
 export const ReactExtension: ExtensionManifest = {
   name: 'react',
-  dependencies: [BuildExt],
-  provider: async (config: {}, [build]: [Build]) => {
-    build.registerTask('react', reactTask);
+  dependencies: [PipesExt],
+  provider: async (config: {}, [pipes]: [Pipes]) => {
+    pipes.registerTask('react', reactTask);
   }
 };
